+++ src/main/java/com/example/saucedemo/framework/pageobject/CheckoutCompletePage.java
@@ -8,7 +8,7 @@ import org.openqa.selenium.WebDriver;
 @Slf4j
 public class CheckoutCompletePage extends BasePage implements PageLoadable<CheckoutCompletePage> {
 
-  private final By confirmationMessage = By.cssSelector("[data-test='complete-header']");
+  private static final By CONFIRMATION_MESSAGE = By.cssSelector("[data-test='complete-header']");
 
   public CheckoutCompletePage(WebDriver driver) {
     super(driver);
@@ -17,13 +17,13 @@ public class CheckoutCompletePage extends BasePage implements PageLoadable<Check
   @Override
   public CheckoutCompletePage waitUntilLoaded() {
     waitUntilUrlContains("checkout-complete");
-    waitUtils.waitUntilVisible(confirmationMessage);
+    waitUtils.waitUntilVisible(CONFIRMATION_MESSAGE);
     return this;
   }
 
   @Step("Get checkout confirmation message")
   public String getConfirmationMessage() {
     log.info("Retrieving checkout confirmation message");
-    return waitUtils.waitUntilVisible(confirmationMessage).getText();
+    return waitUtils.waitUntilVisible(CONFIRMATION_MESSAGE).getText();
   }
 }